+++ src/modules/vga/textmode_ops.cpp
@@ -76,7 +76,7 @@ size_t text_seek(vga_instance *inst, size_t pos, uint32_t flags){
             cpos=pos;
         }
         if(cpos > current_mode->vidmode.height * current_mode->vidmode.width){
-            dbgpf("VGA: Bad text seek: %i\n", cpos);
+            dbgpf("VGA: Bad text seek: %i\n", (int)cpos);
             cpos=current_mode->vidmode.height * current_mode->vidmode.width;
         }
         text_row=cpos/current_mode->vidmode.width;
